timestamp: Rename key resource (#8)
* timestamp: Rename key resource

Plan is to setup timestamp with a cert chain that contains
a self signed root certificate and a signing certificate.

Rename the key resource so it does not claim to be "intermediate".

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

* timestamp: Change value of key encryption key name

This key has been created in staging KMS once already. I have not been
able to confirm if re-creating destroyed keys is possible: this should
avoid any complications.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>

---------

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>