Add LB and URL map resources for tiles_tlog (#24)
This change adds a load balancer with a CDN and SSL for tiles_tlog.
Since the load balancer must conditionally route to either the
Kubernetes service or a bucket, it can't be managed in Kubernetes by the
Ingress resource and instead must be managed explicitly in Terraform.
The Helm chart will use annotations on the Service to create Network
Endpoint Groups[1] to which the load balancer can route.

Since the NEGs are created by Kubernetes, they won't exist the first time these
Terraform modules are run for a given shard. The order of operations on
first deployment of any shard is:

1. Run Terraform with network_endpoint_group_zones in the tiles-tlog
   module set to empty.
2. Update the ArgoCD configuration to deploy the new tlog shard.
3. Update the Terraform input to set network_endpoint_group_zones to the
   all the zones for the GKE cluster's region.

[1] https://cloud.google.com/load-balancing/docs/negs/zonal-neg-concepts

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>