Change gRPC healthcheck to test a TCP connection
gRPC healthchecks on GCP fail with connection issues regarding IPV6. I'm
not sure what the root cause is, but I have found documentation that
gRPC is primarily for Cloud Mesh Service and that gRPC healthchecks
don't support TLS, which we have enabled for routing between the load
balancer and backend service. The documentation suggests just using TCP,
which only checks for a valid connection.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>