Add monitoring for Rekor v2 (#64)
Copy the relevant alerts from Rekor v1 to a new Rekor v2 submodule in
monitoring.

Like the main tiles_tlog module, this module may be re-used for
individual shards, so it is not included in the monitoring module's
sigstore.tf composition and instead must be individually called at the
same level the tiles_tlog module is called.

The module is named rekorv2 instead of tiles_tlog because some of the
metrics are specific to the rekor-tiles service and would not be
relevant to a generic transparency log. This is also why it must be
called on its own and is not included as part of the tiles_tlog module.

There is no uptime check configured for the service itself because Rekor
v2 has no native publicly exposed GET endpoint and it would not be an
effective use of the log to POST entries to it every 30 seconds. The
prober will check that the service is up and working, and the Kubernetes
alerts should notify of the service behaving badly.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>