Set connection drain timeout for tiles_tlog
Connection draining[1] is recommended gives the external load balancer a
grace period to drain connections from the application. The default
timeout set by terraform is 300 seconds[2]. This is excessive
considering the default termination grace period for a Kubernetes pod is
30 seconds, and a healthy Rekor request should last no
more than 10 seconds. Set the timeout to 15 seconds to allow for a
reasonably short termination period. The Kubernetes termination timeout
will be updated as well.

[1] https://cloud.google.com/kubernetes-engine/docs/how-to/ingress-configuration#500-series-errors
[2] https://github.com/hashicorp/terraform-provider-google/issues/13478

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>