Append random ID to Tessera bucket name (#83)
Read traffic will be served from GCS through a CDN. The CDN requires
that the bucket be public. Currently, the bucket name is predictable so
the CDN could be bypassed. This makes the bucket name unguessable. For
backwards compatibility, the ID length is configurable since we have two
buckets in staging without random IDs.

Ref https://github.com/sigstore/rekor-tiles/issues/355

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>
Co-authored-by: Hayden <8418760+haydentherapper@users.noreply.github.com>