Add Cloud Armor rules for HTTP, gRPC and GCS backends (#87)
* Add Cloud Armor rules for HTTP, gRPC and GCS backends

This adds rate limiting for read and write requests for tile-backed
logs. This also restricts request sizes.

The defaults are reasonable, and variables have been provided to tweak
rate limiting to meet deployment needs.

We've chosen to add rate limiting for the read path to prevent unlimited
reads, which come with egress costs. The threshold is much higher than
writes to support monitoring, and will be increased over time for the
public deployment as write QPS increases.

Ref https://github.com/sigstore/rekor-tiles/issues/354

Ref https://github.com/sigstore/rekor-tiles/issues/355

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>

* Remove read body limit, apply json parsing to specific content types

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>

* Update gcp/modules/tiles_tlog/network.tf

Co-authored-by: Bob Callaway <bobcallaway@users.noreply.github.com>
Signed-off-by: Hayden <haydentherapper@users.noreply.github.com>

---------

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>
Signed-off-by: Hayden <haydentherapper@users.noreply.github.com>
Co-authored-by: Hayden <8418760+haydentherapper@users.noreply.github.com>
Co-authored-by: Bob Callaway <bobcallaway@users.noreply.github.com>