monitoring: revert tweak to user-agent regex in tuf metric (#130)
This
 * reverts commit 0dbf4c3dd319f16c2ca7c1166cf5486a0c4cd46e
 * stops trying to accomodate user agent names with whitespace

We cannot have nice things: GitHub uses user-agents of style
    GitHub CLI 00e0c48
    GitHub CLI 2.74.2
and it does so at scale leading to hundreds of new user agents per week
making the metrics less useful.

This is incompatible with relevant RFCs and impossible to parse in a
useful way. Start using whitespace as delimiter as the specs say so all
of these UAs should just show up as unversioned "GitHub".

After this the only non-standard thing in the regex is stripping "v"
from version numbers (something many of our go projects like to do).

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>