Fix backend bucket resource name
Undo part of f2546ab3 to avoid renaming this resource. Terraform plan
shows that renaming it would destroy it which terraform will not allow
us to do and which would cause downtime if it was possible. The change
was unnecessary anyway, the backend_bucket_suffix string should be
enough to distinguish different resources when they are instantiated for
rekor or for ctlog.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>